INT-4208: Support for AsycRestTemplate and thus noblocking IO https://jira.spring.io/browse/INT-4208
Add AsyncRestTemplate support in HttpRequestExecutingMessageHandler and corresponding unit tests
