+++ spring-integration-http/src/main/java/org/springframework/integration/http/config/HttpOutboundGatewayParser.java
@@ -16,6 +16,7 @@
 
 package org.springframework.integration.http.config;
 
+import org.springframework.integration.http.outbound.AsyncHttpRequestExecutingMessageHandler;
 import org.w3c.dom.Element;
 
 import org.springframework.beans.factory.support.BeanDefinitionBuilder;
@@ -42,7 +43,14 @@ public class HttpOutboundGatewayParser extends AbstractConsumerEndpointParser {
 
 	@Override
 	protected BeanDefinitionBuilder parseHandler(Element element, ParserContext parserContext) {
-		BeanDefinitionBuilder builder = BeanDefinitionBuilder.genericBeanDefinition(HttpRequestExecutingMessageHandler.class);
+		BeanDefinitionBuilder builder;
+		boolean async = element.getLocalName().contains("async");
+		if (async) {
+			builder = BeanDefinitionBuilder.genericBeanDefinition(AsyncHttpRequestExecutingMessageHandler.class);
+		}
+		else {
+			builder = BeanDefinitionBuilder.genericBeanDefinition(HttpRequestExecutingMessageHandler.class);
+		}
 
 		HttpAdapterParsingUtils.configureUrlConstructorArg(element, parserContext, builder);
